some more outlining of other services/products

+++ src/Plaid.Net/Transactions/TransactionLocation.cs
@@ -12,10 +12,10 @@ namespace Plaid.Net.Transactions
         
         public string Country { get; set; }
         
-        public decimal Lat { get; set; }
+        public double Lat { get; set; }
         
-        public decimal Lon { get; set; }
+        public double Lon { get; set; }
         
-        public int StoreNumber { get; set; }
+        public long StoreNumber { get; set; }
     }
 }
